feat:encapsulation comment component
